EMI-ats WIP.
Contributors: Fengming Yuan, Joe Beisman, and Ethan Coon, ORNL.

This adds a new ELM external model, ATS, that provides an alternative
way to solve for surface and subsurface flow.  Eventually it will be
extended with energy and chemical component transport.

This is a rebase/squash of all commits used in developing that
capability through Jan 15, 2025.  More commits will be coming, it is
not correct or complete yet.